+++ examples/base_example.rs
@@ -1,15 +1,15 @@
 use actix_web::dev::ServiceRequest;
 use actix_web::{get, middleware, web, App, Error, HttpResponse, HttpServer};
 
-use actix_web_grants::authorities::{AuthDetails, AuthoritiesCheck};
-use actix_web_grants::{proc_macro::has_authorities, AuthorityGuard, GrantsMiddleware};
+use actix_web_grants::permissions::{AuthDetails, PermissionsCheck};
+use actix_web_grants::{proc_macro::has_permissions, GrantsMiddleware, PermissionGuard};
 
 const ROLE_ADMIN: &str = "ROLE_ADMIN";
 const ADMIN_RESPONSE: &str = "Hello Admin!";
 const OTHER_RESPONSE: &str = "Hello!";
 
 #[get("/admin")]
-#[has_authorities("ROLE_ADMIN")]
+#[has_permissions("OP_READ_ADMIN_INFO")]
 // An example of protection via `proc-macro`
 async fn macro_secured() -> HttpResponse {
     HttpResponse::Ok().body(ADMIN_RESPONSE)
@@ -18,7 +18,7 @@ async fn macro_secured() -> HttpResponse {
 #[get("/")]
 // An example of programmable protection
 async fn manual_secure(details: AuthDetails) -> HttpResponse {
-    if details.has_authority(ROLE_ADMIN) {
+    if details.has_permission(ROLE_ADMIN) {
         return HttpResponse::Ok().body(ADMIN_RESPONSE);
     }
     HttpResponse::Ok().body(OTHER_RESPONSE)
@@ -38,7 +38,7 @@ async fn main() -> std::io::Result<()> {
             .service(
                 web::resource("/guard_admin")
                     .to(|| async { HttpResponse::Ok().finish() })
-                    .guard(AuthorityGuard::new(ROLE_ADMIN.to_string())),
+                    .guard(PermissionGuard::new(ROLE_ADMIN.to_string())),
             )
     })
     .bind("localhost:8081")?
@@ -48,7 +48,7 @@ async fn main() -> std::io::Result<()> {
 }
 
 async fn extract(_req: &ServiceRequest) -> Result<Vec<String>, Error> {
-    // Here is a place for your code to get user authorities/grants/permissions from a request
+    // Here is a place for your code to get user permissions/grants/permissions from a request
     // For example from a token or database
 
     // Stub example