+++ tests/permissions_check/manual_check.rs
@@ -3,7 +3,7 @@ use actix_web::{get, test, App, HttpResponse};
 
 use crate::common::{self, ROLE_ADMIN, ROLE_MANAGER};
 use actix_web::http::{header::AUTHORIZATION, StatusCode};
-use actix_web_grants::authorities::{AuthDetails, AuthoritiesCheck};
+use actix_web_grants::permissions::{AuthDetails, PermissionsCheck};
 use actix_web_grants::GrantsMiddleware;
 
 const ADMIN_RESPONSE: &str = "Hello Admin!";
@@ -11,7 +11,7 @@ const OTHER_RESPONSE: &str = "Hello!";
 
 #[get("/")]
 async fn different_body(details: AuthDetails) -> HttpResponse {
-    if details.has_authority(ROLE_ADMIN) {
+    if details.has_permission(ROLE_ADMIN) {
         return HttpResponse::Ok().body(ADMIN_RESPONSE);
     }
     HttpResponse::Ok().body(OTHER_RESPONSE)
@@ -19,7 +19,7 @@ async fn different_body(details: AuthDetails) -> HttpResponse {
 
 #[get("/admin")]
 async fn only_admin(details: AuthDetails) -> HttpResponse {
-    if details.has_authority(ROLE_ADMIN) {
+    if details.has_permission(ROLE_ADMIN) {
         return HttpResponse::Ok().body(ADMIN_RESPONSE);
     }
     HttpResponse::Forbidden().finish()