+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Change crate category to authentication
 
 ## [v1.0.0] - 2021-01-18
 ### Added