+++ CHANGELOG.md
@@ -9,13 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Change crate category to authentication
+- Breaking change(!): change `authoritites` to `permissions` everywhere for more clarity
 
 ## [v1.0.0] - 2021-01-18
 ### Added
 - Github Actions: Add CI pipeline
 
 ### Changed
-- Breaking change(!): remove Arc usage in `AuthoritiesExtractor` [#1](https://github.com/DDtKey/actix-web-grants/pull/1)
+- Breaking change(!): remove Arc usage in `PermissionsExtractor` [#1](https://github.com/DDtKey/actix-web-grants/pull/1)
 
 ## [v0.1.6] - 2021-01-18
 ### Changed